drm/i915: Track whether an atomic transaction changes the active CRTC's
For the purposes of DDB re-allocation we need to know whether a
transaction changes the list of CRTC's that are active.  While
state->modeset could be used for this purpose, that slightly too
aggressive since it would lead us to re-allocate the DDB when a CRTC's
mode changes, but not its active state.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>