drm/i915/gen9: Store pipe DDB entry in crtc state
As we move toward pre-calculating DDB and watermarks during the atomic
check phase, we'll need to know the per-pipe DDB allocation.  In cases
where we're doing a modeset that changes the list of active CRTC's, then
we'll be grabbing all locks and can safely look at the active_crtcs
value to calculate this allocation.  However when the active CRTC list
is not changing, then we won't hold full locks and can't safely use the
active_crtcs value since we might be racing with other modesets that
change it.  However in those cases the pipe's allocation
shouldn't be changing, so there's really no reason to be recalculating
it at all; if we store the pipe DDB entry in the CRTC state that we can
inherit it through atomic commits and only update it on transactions
that actually change it (and hold all the necessary locks).

Note that to make this work properly, we need to ensure we set the
initial value during HW readout at boot, and also ensure that we don't
clobber the old value in clear_intel_crtc_state().

Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>