drm/i915/skl: Read memory characteristics from MCH BAR
There are some upcoming workarounds for SKL-style watermarks that need
to be applied according to system memory characteristics (e.g, only when
using dual channel, single rank memory, or only when exceeding some
percentage of total system memory bandwidth, etc.)  The information we
need is available via the MCH BAR, so read it from there at driver
startup.

Note that this patch only handles Skylake at the moment.  The registers
we need to read may have moved for those platforms, so we'll followup
with additional patches for those platforms in the future once we know
for sure where they are.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>