drm: Introduce detailed plane-blending properties
Some hardware can perform more advanced inter-plane blending than our current
DRM API can express.  E.g., userspace may want to distinguish between
premultiplied vs non-premultiplied alpha for an RGBA framebuffer, may wish to
apply a plane-wide constant alpha factor (multiplied by any per-pixel alpha
already being handled), etc.

We'll model our DRM blending-control properties after the GL's blend state
design.  Drivers will be able to expose 'source factor' and 'destination
factor' properties for planes that expresses how the plane should be blended
with the content below it, and a 'blend color' property will be used to
supply additional color values for the factor combinations that need it.  The
combination of blend factors should be expressive enough that the result of
blending a pixel of a plane (the source) with the pixels of planes below it
(the destination) can be expressed as

	C = SRCFACTOR * SRCCOLOR + DESTFACTOR * DESTCOLOR

Examples (SrcFactor, DestFactor):
 * (ONE, ZERO) --- Show plane content with no blending (i.e., treat as XRGB
	even for ARGB format framebuffers)
 * (ZERO, ONE) --- Invisible plane (pretty useless)
 * (ONE, ONE_MINUS_SRC) --- Blend with per-pixel premultiplied alpha
 * (SRC, ONE_MINUS_SRC) --- Blend with per-pixel non-premultiplied alpha
 * (CONSTANT, ONE_MINUS_CONSTANT) --- XRGB content with fixed
        plane-wide transparency; alpha value of 'blend color'
        property used for global alpha value
 * (CONSTANT, ONE_MINUS_CONSTANT_TIMES_SRC) --- Blend with both
	per-pixel premultiplied alpha and a constant alpha factor
 * (CONSTANT_TIME_SRC, ONE_MINUS_CONSTANT_TIMES_SRC) --- Blend with both
        per-pixel non-premultiplied alpha and a constant alpha factor
 * (AUTO, AUTO) --- Use default (backward-compatible) blending according to the
        XRGB or ARGB format of the plane's framebuffer

These new properties will only be exposed to atomic-capable clients so that we
can be confident that the properties will be updated together.

Cc: dri-devel@lists.freedesktop.org
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>