cgroup: Add notifier call chain for cgroup destruction
Drivers or other kernel subsystems may allow subsystem-specific policy and
configuration to be applied to cgroups.  If these subsystems track private data
on a per-cgroup basis, they need a way to be notified about cgroup destruction
so that they can clean up their own internal data for that cgroup.  Let's add a
blocking_notifier that will be called whenever a cgroup in the v2 hierarchy is
destroyed to allow this cleanup.

I'm arbitrarily restricting this behavior to the default (cgroup-v2) hierarchy
for now since I don't anticipate it being terribly useful on the legacy
hierarchies.  We can certainly relax that restriction if someone comes up with
a use case that needs this on the v1 hierarchies.

Cc: Tejun Heo <tj@kernel.org>
Cc: cgroups@vger.kernel.org
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>