drm: Introduce CRTC gangs
A gang is a pair of CRTCs that a driver uses internally to drive the
configuration that userspace requested on a single CRTC.  Some hardware
allows the output of two CRTC's to be joined into a single image that is
passed to downstream encoders and connectors as if it came from a single
CRTC.

The use of CRTC gangs is a driver/hardware internal detail and should
remain invisible to userspace.  A gang may only be formed when userspace
has left one of the CRTCs in a disabled state; the driver is free to
steal that CRTC and make use of it internally, but there should be no
change to any of the properties or state fields for the CRTC that
userspace believes to be disabled.  The "disabled" CRTC that the driver
steals for internal gang usage is considered the "slave" of the gang,
and the CRTC that userspace provided a configuration for is considered
the "master."

Note that userspace may attempt to start using a gang slave on a
subsequent atomic transaction.  If the driver is not able to transfer
the slave responsibilities to a different disabled CRTC, the new atomic
transaction must be rejected to avoid overcommitting the hardware.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>