drm: Introduce drm_crtc_hw_active()
crtc_state->active reflects the uapi "active" status of a CRTC.  However
with the introduction of CRTC gangs, drivers may wish to quickly check
whether a CRTC is active at the hardware level independently of what we
tell userspace.  Let's add a quick helper to take care of chasing down
the gang master to get the "real" hardware active status.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>