drm/i915/icl: Introduce i915_adjust_gang_planes()
When a uapi-provided mode is serviced by the two CRTCs in a CRTC gang,
we need to grab a bunch of extra planes on the slave CRTC and then
clip the plane rectangles on both CRTCs so that planes are properly
showing the half of the content that their CRTC is responsible for.
Let's introduce a helper that will be called in subsequent patches to
implement big joiner-based modes.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>