drm/i915/icl: Big joiner groundwork
ICL allows us to gang two pipes together to drive a single
transcoder+DDI; this allows us to support display modes that exceed the
clock or resolution limits of a single pipe.  Usage of the big joiner
should remain transparent to userspace; userspace will request a single
mode (e.g., 5k@120) on a single CRTC (the "master"); internally the
driver will find a suitable second CRTC (the "slave") and program both
the master and slave pipes+planes such that each provides half of the
content.

At least on current hardware (ICL), our choices for how to setup the big
joiner are limited -- we can only use pipe B as master and pipe C as
slave.

Let's start setting up the groundwork for big joiner enablement --- we
need to track whether a CRTC is participating in a big joiner config
(and if so whether it's the master or slave) and provide an easy way to
obtain the slave's CRTC state given a master CRTC state.  We'll wait to
actually enable big joiner mode until a much later patch after all the
necessary support throughout the driver is in place.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>